requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.9.23
future==1.0.0
gast==0.6.0
geographiclib==2.1
geojson==3.2.0
geopy==2.4.1
gevent==25.9.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gTTS==2.5.4
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pycparser==2.23
Pygments==2.19.2
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.3
srt==3.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
vosk==0.3.45
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
wsproto==1.2.0
zope.event==6.1
zope.interface==8.1
